refactor: move populateFriendsDB into its own file
Fix the populateFriendsDB function running twice due to how export
works. Export evaluates the file and runs all the invoked functions
inside a module before exporting functions. This causes the function
to run twice inside the populateDB. To fix this move the script
functions inside another file and import it to both populateFriendsDB
and populateDB files.
